Modernize packaging: pyproject + pixi + Python >=3.12; rename bem -> braggedgemodeling
Foundation of the modernization (EWM story). Brings the package up to the
neutron-imaging fleet standard, modeled on iBeatles.

- Rename package bem/ -> braggedgemodeling/ (history-preserving); update all
  imports. PyPI distribution name will be 'braggedgemodeling' ('bem' is taken).
  A README note (follow-up docs task) will give 'import braggedgemodeling as bem'
  as a drop-in stop-gap. Runtime config file name 'bem.conf' kept for back-compat.
- setup.py + hand-maintained _version.py -> pyproject.toml (PEP 621) + setuptools
  + versioningit; __init__ reads the version via importlib.metadata.
- Adopt pixi (conda-forge): test/py312/py313/package/docs features; tasks test,
  build (PyPI), build-conda (conda mambabuild, conda-forge+neutronimaging).
- conda.recipe/meta.yaml (noarch python) for the conda build.
- Python >=3.12 (SPEC 0); CI replaced with a pixi unit-test matrix (py312, py313),
  SHA-pinned actions. Removed Travis config.
- Texture subpackage (MATLAB/mtex) kept as an optional extra; its tests are
  skipped in CI (see #40).

Verified locally (pixi run test): 19 passed on the modernized stack.
Grype/Dependabot and PyPI/conda publishing land in follow-up PRs (publishing
needs the joint PyPI trusted-publisher setup).

Assisted-With: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
