+++ .github/workflows/CI.yml
@@ -1,54 +1,35 @@
 name: CI
 
-env: {}
+on:
+  push:
+    branches: [main, master]
+  pull_request:
+  workflow_dispatch:
 
 permissions:
   contents: read
 
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-  schedule:
-    - cron: "30 7 * * *"
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
 jobs:
-  build-and-test:
-    runs-on: ${{ matrix.os }}
+  unit-test:
+    name: Unit tests (${{ matrix.environment }})
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest"]
-        python-version: ["3.8"]  
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
+        environment: [py312, py313]
     steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v2
-
-      # setup conda
-      - uses: conda-incubator/setup-miniconda@9f54435e0e72c53962ee863144e47a4b094bfd35 # v2
+      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
         with:
-          auto-update-conda: true
-          activate-environment: testenv
-          python-version: ${{ matrix.python-version }}
-          channels: conda-forge
-
-      # install deps and build
-      - name: install / build
-        shell: pwsh
-        run: |
-          conda install pip pytest pytest-cov pyyaml numpy scipy matplotlib pandas
-          conda install -c diffpy diffpy.Structure
-          conda install periodictable
-          python setup.py install
-      # test
-      - name: test
-        shell: pwsh
-        run: |
-          py.test --cov=bem tests
-          ./tests/Test_xs_NaCl.py noplot
-          
+          fetch-depth: 0
+          fetch-tags: true
+      - uses: prefix-dev/setup-pixi@a09b6247153796b190642a2b53fac4241043cf6f # v0.10.0
+        with:
+          manifest-path: pyproject.toml
+          environments: ${{ matrix.environment }}
+      - name: Run unit tests
+        run: pixi run -e ${{ matrix.environment }} test --cov=braggedgemodeling --cov-report=xml --cov-report=term
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@fb8b3582c8e4def4969c97caa2f19720cb33a72f # v5
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+        continue-on-error: true
