--- .travis.yml
@@ -1,57 +1,0 @@
-sudo: false
-# dist: trusty
-
-language: python
-
-python:
-  - 2.7
-  - 3.6
-
-branches:
-  only:
-    - master
-    - texture
-#
-# services:
-#
-# matrix:
-#  include:
-#    - os: linux
-
-env:
-   
-before_install:
-  - if [ ${TRAVIS_PYTHON_VERSION:0:1} == "2" ]; then wget http://repo.continuum.io/miniconda/Miniconda-3.5.5-Linux-x86_64.sh -O miniconda.sh; else wget http://repo.continuum.io/miniconda/Miniconda3-3.5.5-Linux-x86_64.sh -O miniconda.sh; fi
-  - chmod +x miniconda.sh
-  - ./miniconda.sh -b -p /home/travis/mc
-  - export PATH=/home/travis/mc/bin:$PATH
-  - "export DISPLAY=:99.0" # configure a headless display to test plot generation
-  - "sh -e /etc/init.d/xvfb start"
-  - sleep 3 # give xvfb some time to start
-
-install:
-  - export GIT_FULL_HASH=`git rev-parse HEAD`
-  - conda config --set always_yes true
-  - conda update conda
-  - conda config --add channels conda-forge
-  - conda create -n testenv pip pytest pytest-cov python=$TRAVIS_PYTHON_VERSION
-  - conda install -n testenv pyyaml numpy scipy matplotlib pandas
-  - conda install -n testenv -c diffpy diffpy.Structure
-  - conda install -n testenv periodictable
-  - source activate testenv
-  - python setup.py install
-  - pip install coveralls
-#  - pip install codecov
-
-script:
-#   - travis_wait 60 py.test --cov=bem tests
-   - py.test --cov=bem tests
-   - ./tests/Test_xs_NaCl.py noplot
-   
-after_success:
-   - coveralls
-
-notifications:
-  email:
-    on_success: change
-    on_failure: always
