+++ braggedgemodeling/constants.py
