+++ braggedgemodeling/texture/polfig2VPSC.m
