+++ docs/tutorial_Ni.py
@@ -1,4 +1,4 @@
-from bem.matter import Atom, Lattice, Structure
+from braggedgemodeling.matter import Atom, Lattice, Structure
 atoms = [Atom('Ni', (0,0,0)), Atom('Ni', (0.5, 0.5, 0)),
          Atom('Ni', (0.5,0,0.5)), Atom('Ni', (0, 0.5, 0.5))]
 a=3.5238
@@ -11,7 +11,7 @@ import numpy as np
 wavelengths = np.arange(0.05, 5.5, 0.005)
 T = 300
 # create calculator
-from bem import xscalc
+from braggedgemodeling import xscalc
 xscalculator = xscalc.XSCalculator(fccNi, T, max_diffraction_index=7)
 # compute various contributions
 # In neutron Bragg Edge data analysis, it may not be necessary to calculate all these
@@ -37,7 +37,7 @@ plt.legend()
 plt.show()
 
 # texture
-from bem import xtaloriprobmodel as xopm
+from braggedgemodeling import xtaloriprobmodel as xopm
 texture_model = xopm.MarchDollase()
 texture_model.r[(0,0,1)] = 2
 xscalculator = xscalc.XSCalculator(fccNi, T, texture_model)
@@ -46,7 +46,7 @@ plt.show()
 
 
 # peak profile
-from bem import peak_profile as pp, calc
+from braggedgemodeling import peak_profile as pp, calc
 jorgensen = pp.Jorgensen(alpha=[50, 0.], beta=[10, 0], sigma=[0, .003, 0])
 spectrum_calculator = calc.BraggEdgeSpectrumCalculator(xscalculator, jorgensen)  
 spectrum = spectrum_calculator('total', wavelengths)
