+++ pyproject.toml
@@ -1,0 +1,123 @@
+[project]
+name = "braggedgemodeling"
+description = "Bragg Edge modeling"
+dynamic = ["version"]
+requires-python = ">=3.12"
+license = { text = "BSD" }
+keywords = ["instrument", "neutron", "imaging", "Bragg Edge"]
+readme = "README.md"
+authors = [{ name = "ORNL neutron imaging team", email = "neutronimaging@ornl.gov" }]
+maintainers = [{ name = "ORNL neutron imaging team", email = "neutronimaging@ornl.gov" }]
+dependencies = [
+  "pyyaml",
+  "numpy",
+  "scipy",
+  "matplotlib",
+  "diffpy.structure",
+  "periodictable",
+  "pandas",
+]
+
+# The texture subpackage additionally requires an external MATLAB + mtex + VPSC
+# toolchain (not installable from PyPI/conda). See docs; tracked in issue #40.
+[project.optional-dependencies]
+texture = []
+
+[project.urls]
+homepage = "https://github.com/ornlneutronimaging/braggedgemodeling"
+repository = "https://github.com/ornlneutronimaging/braggedgemodeling"
+issues = "https://github.com/ornlneutronimaging/braggedgemodeling/issues"
+
+[build-system]
+requires = ["setuptools>=61", "versioningit"]
+build-backend = "setuptools.build_meta"
+
+[tool.versioningit.vcs]
+method = "git"
+default-tag = "0.1.3"
+
+[tool.versioningit.next-version]
+method = "minor"
+
+[tool.versioningit.format]
+distance = "{next_version}.dev{distance}+{vcs}{rev}"
+dirty = "{next_version}.dev{distance}+{vcs}{rev}.dirty"
+distance-dirty = "{next_version}.dev{distance}+{vcs}{rev}.dirty"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["braggedgemodeling*"]
+exclude = ["tests*", "docs*", "scripts*"]
+
+[tool.setuptools.package-data]
+# ship the MATLAB helpers used by the texture backend
+"braggedgemodeling.texture" = ["*.m"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+# texture tests need MATLAB/mtex, which is unavailable in CI (see issue #40)
+markers = ["matlab: tests that require an external MATLAB + mtex install"]
+
+# ===================
+# ==== Pixi ==========
+# ===================
+
+[tool.pixi.workspace]
+channels = ["conda-forge"]
+platforms = ["linux-64", "osx-arm64"]
+preview = []
+
+[tool.pixi.pypi-dependencies]
+braggedgemodeling = { path = ".", editable = true }
+
+[tool.pixi.dependencies]
+python = ">=3.12"
+pyyaml = "*"
+numpy = "*"
+scipy = "*"
+matplotlib-base = "*"
+pandas = "*"
+periodictable = "*"
+"diffpy.structure" = "*"
+
+[tool.pixi.tasks]
+test = { cmd = "pytest --ignore=tests/texture --ignore=tests/test_texture.py --ignore=tests/test_texture_Fe.py", description = "Run the MATLAB-free test suite (texture tests need MATLAB/mtex — see issue #40)" }
+build = { cmd = "python -m build", description = "Build the PyPI sdist + wheel" }
+build-conda = { cmd = "cd conda.recipe && VERSION=$(versioningit ../) conda mambabuild --channel conda-forge --channel neutronimaging --output-folder . .", description = "Build the conda package" }
+
+[tool.pixi.feature.test.dependencies]
+pytest = ">=7"
+pytest-cov = "*"
+
+[tool.pixi.feature.package.dependencies]
+python-build = "*"
+versioningit = "*"
+boa = "*"
+conda-build = "*"
+anaconda-client = "*"
+
+[tool.pixi.feature.docs.dependencies]
+sphinx = "*"
+sphinx_rtd_theme = "*"
+
+[tool.pixi.feature.py312.dependencies]
+python = "3.12.*"
+
+[tool.pixi.feature.py313.dependencies]
+python = "3.13.*"
+
+[tool.pixi.environments]
+default = { features = ["test"], solve-group = "default" }
+docs = { features = ["docs"], solve-group = "default" }
+# package build isolated in its own solve-group (boa/conda-build pins)
+package = { features = ["package"], no-default-feature = true }
+py312 = { features = ["test", "py312"], solve-group = "py312" }
+py313 = { features = ["test", "py313"], solve-group = "py313" }
+
+[tool.ruff]
+line-length = 120
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "W"]
+
+[tool.ruff.format]
