+++ scripts/vdrive_handler_to_mtex.py
@@ -1,7 +1,7 @@
 import argparse
 
-from bem.texture.preparation.vdrive_handler import VDriveHandler
-from bem.texture.preparation.vdrive_to_mtex import VDriveToMtex
+from braggedgemodeling.texture.preparation.vdrive_handler import VDriveHandler
+from braggedgemodeling.texture.preparation.vdrive_to_mtex import VDriveToMtex
 
 parser = argparse.ArgumentParser(description='VDrive File Handler')
 parser.add_argument('-i', '--input', help='VDrive file', type=str)
