+++ tests/test_extinction.py
@@ -6,8 +6,8 @@ from __future__ import print_function
 interactive = False
 
 import numpy as np
-from bem import xscalc, diffraction
-from bem.matter import fccNi
+from braggedgemodeling import xscalc, diffraction
+from braggedgemodeling.matter import fccNi
 
 def test_fccNi():
     lambdas = np.arange(0.05, 5.5, 0.001)
