+++ tests/test_xs_Al.py
@@ -4,8 +4,8 @@
 interactive = False
 
 import os, numpy as np
-from bem import xscalc, diffraction
-from bem.matter import fccAl
+from braggedgemodeling import xscalc, diffraction
+from braggedgemodeling.matter import fccAl
 
 thisdir = os.path.dirname(__file__)
