+++ tests/texture/test_workflow.py
@@ -13,7 +13,7 @@ rpffile = 'vulcan.rpf'
 vpscfile = 'vpsc.txt'
 hkls = [[1,1,1], [2,0,0], [2,2,0], [2,2,2]]
 Rsamples_file = 'Rsamples.dat'
-from bem import matter
+from braggedgemodeling import matter
 # FCC
 atoms = [matter.Atom('Ni', (0,0,0)), matter.Atom('Ni', (0.5, 0.5, 0)),
          matter.Atom('Ni', (0.5,0,0.5)), matter.Atom('Ni', (0, 0.5, 0.5))]
@@ -26,8 +26,8 @@ fccNi = matter.Structure(atoms, lattice, sgid=225)
 
 def test1():
     # vdrive_filename.txt -> vulcan.rpf
-    from bem.texture.preparation.vdrive_handler import VDriveHandler
-    from bem.texture.preparation.vdrive_to_mtex import VDriveToMtex
+    from braggedgemodeling.texture.preparation.vdrive_handler import VDriveHandler
+    from braggedgemodeling.texture.preparation.vdrive_to_mtex import VDriveToMtex
     o_vdrive = VDriveHandler(filename = infile)
     o_vdrive.run()
     o_vdrive.export(filename = interm)
@@ -39,14 +39,14 @@ def test1():
 
 def test2():
     # vulcan.rpf -> vpsc.txt
-    from bem.texture import mtex
+    from braggedgemodeling.texture import mtex
     mtex.setup()
     mtex.polfig2VPSC(rpffile, vpscfile, hkls, Npoints=20000)
     return
 
 def test3():
     # texture -> R
-    from bem.texture import texture2R
+    from braggedgemodeling.texture import texture2R
     return texture2R.compute(
         fccNi,
         tex = vpscfile,
@@ -60,14 +60,14 @@ def test4():
     # cross section from R
     wavelengths = np.arange(0.05, 5.5, 0.001)
     T = 300
-    from bem import xscalc
+    from braggedgemodeling import xscalc
     calc = xscalc.XSCalculator(fccNi, T, max_diffraction_index=3)
     coh_el_xs = [calc.xs_coh_el(l) for l in wavelengths]
     # add texture
-    from bem.texture import texture2R
+    from braggedgemodeling.texture import texture2R
     lambdas, Rs = texture2R.read_results(Rsamples_file)
     hkls = [eval(l) for l in open('hkls.txt').readlines()]
-    from bem.texture.InterpolatedXOPM import InterpolatedXOPM
+    from braggedgemodeling.texture.InterpolatedXOPM import InterpolatedXOPM
     texture_model = InterpolatedXOPM(hkls, lambdas, Rs)
     calc2 = xscalc.XSCalculator(fccNi, T, texture_model, max_diffraction_index=3)
     coh_el_xs2 = [calc2.xs_coh_el(l) for l in wavelengths]
